Create generic Button component, rename ButtonLink to Button
